# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Сервис динамической сегментации пользователей",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/segments": {
            "get": {
                "description": "Получает все существующие сегменты",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "segment"
                ],
                "summary": "Получить все сегменты",
                "responses": {
                    "200": {
                        "description": "Успешная операция",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера при получении сегментов",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorDTO"
                        }
                    }
                }
            },
            "post": {
                "description": "Добавляет сегмент",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "segment"
                ],
                "summary": "Добавить сегмент",
                "parameters": [
                    {
                        "description": "Параметры для добавления сегмента",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.SegmentDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешная операция",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Невалидный запрос или конфликт",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorDTO"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorDTO"
                        }
                    }
                }
            }
        },
        "/segments/history": {
            "get": {
                "description": "Получает историю добавления/удаления сегментов у пользователя. Если указаны параметры 'year' и 'month', возвращает данные за период. Иначе возвращает всю доступную историю.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "segment"
                ],
                "summary": "Получить историю операций с сегментами",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Год для фильтрации истории (например, 2024)",
                        "name": "year",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Месяц для фильтрации истории (1-12)",
                        "name": "month",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешная операция. Возвращает список записей истории.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.HistoryTableDTO"
                            }
                        }
                    },
                    "400": {
                        "description": "Невалидный запрос: неверный формат года/месяца или невалидный период.",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorDTO"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorDTO"
                        }
                    }
                }
            }
        },
        "/user/{user_id}": {
            "get": {
                "description": "Получает активные сегменты пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Получить сегменты пользователя",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID пользователя",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список активных сегментов пользователя",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.SegmentUserDataDTO"
                            }
                        }
                    },
                    "400": {
                        "description": "Невалидный ID пользователя",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorDTO"
                        }
                    },
                    "404": {
                        "description": "Пользователь не найден",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorDTO"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorDTO"
                        }
                    }
                }
            },
            "post": {
                "description": "Добавляет пользователя к существующему сегменту",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Добавить пользователя к сегменту",
                "parameters": [
                    {
                        "description": "Параметры для добавления/удаления сегментов",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.SegmentDTO"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "ID пользователя",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешная операция",
                        "schema": {
                            "$ref": "#/definitions/model.UserResponseDTO"
                        }
                    },
                    "400": {
                        "description": "Невалидный запрос",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorDTO"
                        }
                    }
                }
            },
            "patch": {
                "description": "Добавляет или удаляет сегменты у пользователя, опционально устанавливая TTL.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Обновить сегменты пользователю",
                "parameters": [
                    {
                        "description": "Параметры для добавления/удаления сегментов у пользователя",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.SegmentUpdateUserDTO"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "ID пользователя",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Операция прошла успешно",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Невалидный запрос или конфликт",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorDTO"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.ErrorDTO": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "model.HistoryTableDTO": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "operation": {
                    "type": "string"
                },
                "segment_slug": {
                    "type": "string"
                },
                "user_ID": {
                    "type": "integer"
                }
            }
        },
        "model.SegmentDTO": {
            "type": "object",
            "properties": {
                "auto_percent": {
                    "type": "integer"
                },
                "slug": {
                    "type": "string"
                }
            }
        },
        "model.SegmentUpdateUserDTO": {
            "type": "object",
            "properties": {
                "addslugs": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "removeslugs": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "ttl_hours": {
                    "type": "integer"
                }
            }
        },
        "model.SegmentUserDataDTO": {
            "type": "object",
            "properties": {
                "autoPercent": {
                    "description": "Процент для автоматического назначения (из segments)",
                    "type": "integer"
                },
                "expiresAt": {
                    "type": "string"
                },
                "isManuallyAssigned": {
                    "description": "Эти поля будут NULL/false, если пользователь не состоит в сегменте вручную",
                    "type": "boolean"
                },
                "slug": {
                    "description": "Имя сегмента (из segments)",
                    "type": "string"
                }
            }
        },
        "model.UserResponseDTO": {
            "type": "object",
            "properties": {
                "received": {
                    "type": "boolean"
                },
                "slug": {
                    "type": "string"
                },
                "userID": {
                    "type": "integer",
                    "format": "int64"
                }
            }
        }
    }
}